refactor(chat): host the Chat SDK gateway in the party worker
Everything already routes through the party relay, so put the Chat SDK there
directly instead of in apps/web. Removes the inter-service hop and the extra
shared secret, and gets the heavy chat/discord.js deps out of the marketing
site (which is what Vercel builds).

- apps/party: Chat SDK bot + adapters live here; the worker fetch routes
  /api/webhooks/* (and /api/discord/gateway) to the bot, everything else to the
  partyserver router. The handler relays to the desktop via the worker's own
  Durable Object (getServerByName) and posts the reply through the live thread.
  wrangler `alias` stubs the Discord adapter's optional native deps.
- apps/web: reverted — chat routes, server/chat, vite alias, env vars and deps
  all removed; back to the plain marketing site.
- desktop + @repo/dispatch wire protocol unchanged.
- Also relay to a single registered device (multiple hosts would each run the
  turn and answer the same correlationId).
- docs/.dev.vars: webhook URLs + all secrets now on the party worker.
