+++ apps/party/.dev.vars.example
@@ -1,6 +1,33 @@
-# Party relay secrets for `wrangler dev` (copy to .dev.vars; never commit).
-# In production: `wrangler secret put CHAT_RELAY_SECRET`.
+# Party relay + chat gateway secrets for `wrangler dev` (copy to .dev.vars;
+# never commit). In production set each with `wrangler secret put <NAME>`.
 
-# Shared secret the chat gateway (@repo/web) must present on inbound POSTs.
-# Must match CHAT_RELAY_SECRET in apps/web. Leave empty to disable the check.
+# --- Bridge ---
+# Shared secret. Must match the desktop's DISPATCH_CHAT_SECRET. The desktop
+# presents it to register as the agent host; leave empty to disable the check
+# (local dev only — set it in production).
 CHAT_RELAY_SECRET=""
+# The desktop's dispatch room code (shown in the Inteligir desktop app).
+CHAT_RELAY_ROOM=""
+# Bot display name across platforms (optional).
+CHAT_BOT_USERNAME="inteligir"
+
+# --- Slack (Events API webhook) ---
+SLACK_BOT_TOKEN=""
+SLACK_SIGNING_SECRET=""
+
+# --- Telegram (Bot API webhook) ---
+TELEGRAM_BOT_TOKEN=""
+TELEGRAM_WEBHOOK_SECRET_TOKEN=""
+
+# --- WhatsApp (Meta Cloud API) ---
+WHATSAPP_ACCESS_TOKEN=""
+WHATSAPP_APP_SECRET=""
+WHATSAPP_PHONE_NUMBER_ID=""
+WHATSAPP_VERIFY_TOKEN=""
+
+# --- Discord (Interactions endpoint; messages need the Gateway — see docs) ---
+DISCORD_BOT_TOKEN=""
+DISCORD_PUBLIC_KEY=""
+DISCORD_APPLICATION_ID=""
+# Bearer secret guarding the /api/discord/gateway listener route.
+CRON_SECRET=""
