+++ apps/party/package.json
@@ -8,8 +8,16 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@chat-adapter/discord": "^4.30.0",
+    "@chat-adapter/slack": "^4.30.0",
+    "@chat-adapter/state-memory": "^4.30.0",
+    "@chat-adapter/telegram": "^4.30.0",
+    "@chat-adapter/whatsapp": "^4.30.0",
     "@repo/dispatch": "workspace:*",
-    "partyserver": "^0.5.6"
+    "ai": "^6.0.182",
+    "chat": "^4.30.0",
+    "partyserver": "^0.5.6",
+    "zod": "^4.0.0"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20260522.1",
