+++ apps/party/src/bot.ts
@@ -1,0 +1,128 @@
+import { Chat, type Message, type Thread } from "chat";
+import { createSlackAdapter } from "@chat-adapter/slack";
+import { createTelegramAdapter } from "@chat-adapter/telegram";
+import { createWhatsAppAdapter } from "@chat-adapter/whatsapp";
+import { createDiscordAdapter } from "@chat-adapter/discord";
+import { createMemoryState } from "@chat-adapter/state-memory";
+import { getServerByName } from "partyserver";
+import type { ChatConversation } from "@repo/dispatch";
+
+import type { Env } from "./env";
+
+// ---------------------------------------------------------------------------
+// Chat SDK gateway — hosted in the party worker so the public webhook endpoint
+// and the desktop relay live in one place (no extra service / inter-service
+// hop). Each adapter normalizes its platform's webhooks into the same
+// Thread/Message shape; a single handler relays the message to the desktop
+// agent through this worker's Durable Object and posts the reply back.
+// ---------------------------------------------------------------------------
+
+const FALLBACK = "Inteligir is offline right now — please try again in a moment.";
+
+let cached: Chat | null = null;
+
+function create(env: Env): Chat {
+  const bot = new Chat({
+    userName: env.CHAT_BOT_USERNAME || "inteligir",
+    adapters: {
+      slack: createSlackAdapter({
+        botToken: env.SLACK_BOT_TOKEN,
+        signingSecret: env.SLACK_SIGNING_SECRET,
+      }),
+      telegram: createTelegramAdapter({
+        botToken: env.TELEGRAM_BOT_TOKEN,
+        secretToken: env.TELEGRAM_WEBHOOK_SECRET_TOKEN,
+        mode: "webhook",
+      }),
+      whatsapp: createWhatsAppAdapter({
+        accessToken: env.WHATSAPP_ACCESS_TOKEN,
+        appSecret: env.WHATSAPP_APP_SECRET,
+        phoneNumberId: env.WHATSAPP_PHONE_NUMBER_ID,
+        verifyToken: env.WHATSAPP_VERIFY_TOKEN,
+      }),
+      discord: createDiscordAdapter({
+        botToken: env.DISCORD_BOT_TOKEN,
+        publicKey: env.DISCORD_PUBLIC_KEY,
+        applicationId: env.DISCORD_APPLICATION_ID,
+      }),
+    },
+    state: createMemoryState(),
+  });
+
+  const handle = makeHandler(env);
+  bot.onDirectMessage(handle);
+  bot.onNewMention(handle);
+  bot.onSubscribedMessage(handle);
+  return bot;
+}
+
+/** Lazily constructed, cached for the lifetime of the isolate. */
+export function getBot(env: Env): Chat {
+  if (!cached) cached = create(env);
+  return cached;
+}
+
+function makeHandler(env: Env) {
+  return async (thread: Thread, message: Message): Promise<void> => {
+    if (message.author.isMe) return;
+    const text = message.text?.trim();
+    if (!text) return;
+
+    const loose = thread as unknown as { id?: string; channelId?: string; isDM?: boolean };
+    const platform = (message as unknown as { adapterName?: string }).adapterName ?? "chat";
+
+    await thread.startTyping().catch(() => {});
+
+    const reply = await relayToDesktop(env, {
+      text,
+      conversation: {
+        platform,
+        channelId: loose.channelId,
+        threadId: loose.id,
+        isDM: loose.isDM,
+        userId: message.author.userId,
+        userName: message.author.userName,
+      },
+    });
+
+    await thread.post(reply ?? FALLBACK);
+  };
+}
+
+/**
+ * Relay one message to the desktop agent via this worker's Durable Object and
+ * wait for the reply. The DO holds the desktop's WebSocket and correlates the
+ * request to the matching `chat_reply`. Returns null when not configured, the
+ * desktop is offline, or the turn times out.
+ */
+async function relayToDesktop(
+  env: Env,
+  input: { text: string; conversation?: ChatConversation },
+): Promise<string | null> {
+  const room = env.CHAT_RELAY_ROOM;
+  if (!room) {
+    console.error("[chat] CHAT_RELAY_ROOM not configured");
+    return null;
+  }
+
+  try {
+    const stub = await getServerByName(env.DispatchServer, room);
+    const res = await stub.fetch("https://dispatch.internal/relay", {
+      method: "POST",
+      headers: {
+        "content-type": "application/json",
+        "x-relay-secret": env.CHAT_RELAY_SECRET ?? "",
+      },
+      body: JSON.stringify({ text: input.text, conversation: input.conversation }),
+    });
+    if (!res.ok) {
+      console.error(`[chat] relay failed: ${res.status}`);
+      return null;
+    }
+    const data = (await res.json()) as { text?: string };
+    return data.text ?? null;
+  } catch (err) {
+    console.error("[chat] relay error:", err);
+    return null;
+  }
+}
