--- apps/web/.dev.vars.example
@@ -1,33 +1,0 @@
-# Chat gateway secrets for `wrangler dev` (copy to .dev.vars; never commit).
-# In production set these with `wrangler secret put <NAME>` (or the dashboard).
-
-# --- Bridge to the desktop agent via the party relay ---
-# Party host without scheme. Dev: "localhost:8787". Prod: "inteligir-party.<acct>.workers.dev".
-CHAT_RELAY_PARTY_HOST="localhost:8787"
-# The desktop's dispatch room code (shown in the Inteligir desktop app).
-CHAT_RELAY_ROOM=""
-# Shared secret; must match the party worker's CHAT_RELAY_SECRET.
-CHAT_RELAY_SECRET=""
-# Bot display name across platforms (optional).
-CHAT_BOT_USERNAME="inteligir"
-
-# --- Slack (Events API webhook) ---
-SLACK_BOT_TOKEN=""
-SLACK_SIGNING_SECRET=""
-
-# --- Telegram (Bot API webhook) ---
-TELEGRAM_BOT_TOKEN=""
-TELEGRAM_WEBHOOK_SECRET_TOKEN=""
-
-# --- WhatsApp (Meta Cloud API) ---
-WHATSAPP_ACCESS_TOKEN=""
-WHATSAPP_APP_SECRET=""
-WHATSAPP_PHONE_NUMBER_ID=""
-WHATSAPP_VERIFY_TOKEN=""
-
-# --- Discord (Interactions endpoint; messages need the Gateway — see docs) ---
-DISCORD_BOT_TOKEN=""
-DISCORD_PUBLIC_KEY=""
-DISCORD_APPLICATION_ID=""
-# Bearer secret guarding the /api/discord/gateway listener route.
-CRON_SECRET=""
